slight change, not need to run 2015 - 2017

+++ Data_Manipulation/ISP_total_population.py
@@ -70,16 +70,19 @@ year = sys.argv[1]
 out_list = list()
 #for year in table_list:
 for state in state_list:
-	query = """
-	SELECT HocoNum, BlockCode ,MAX(HighSpeed) AS HighSpeed
-	FROM {0}
-	WHERE StateAbbr = \"{1}\"
-	AND Consumer = 1
-	GROUP BY HocoNum,BlockCode
-	;
-	""".format(year,state)
-	df = pd.read_sql(query, engine)
-	df.to_csv("/home/analysis/Final/ISP/blocks/{0}/{1}.csv".format(year[:7],state),index=False)
+	if not os.path.exists("/home/analysis/Final/ISP/blocks/{0}/{1}.csv"):
+		query = """
+		SELECT HocoNum, BlockCode ,MAX(HighSpeed) AS HighSpeed
+		FROM {0}
+		WHERE StateAbbr = \"{1}\"
+		AND Consumer = 1
+		GROUP BY HocoNum,BlockCode
+		;
+		""".format(year,state)
+		df = pd.read_sql(query, engine)
+		df.to_csv("/home/analysis/Final/ISP/blocks/{0}/{1}.csv".format(year[:7],state),index=False)
+	else:
+		df = pd.read_csv("/home/analysis/Final/ISP/blocks/{0}/{1}.csv")
 	df['population'] = df['BlockCode'].apply(lambda x: int(ref[state][x]))
 	df = df.groupby(['HocoNum','HighSpeed'])['population'].sum().reset_index()
 	df['state'] = state
